ref: add KeypressMixin for ContactList
